# Optimize update_address_current_token_balances_in_batches