Fixes #1335

## Test Resources

### `count.sql`

Returns the number of rows to update.  This takes 16 minutes on Eth Mainnet Prod.

```sql
SELECT COUNT(*)
FROM (SELECT 1
      FROM address_token_balances
             INNER JOIN address_current_token_balances
                        ON address_current_token_balances.address_hash =
                           address_token_balances.address_hash AND
                           address_current_token_balances.token_contract_address_hash =
                           address_token_balances.token_contract_address_hash
      GROUP BY address_token_balances.address_hash,
               address_token_balances.token_contract_address_hash,
               address_current_token_balances.block_number
      HAVING MAX(address_token_balances.block_number) !=
             address_current_token_balances.block_number) AS address_incorrect_token_balances;
```

### `poison_some.sql`

**DO NOT RUN THIS ON PROD.**

Purposely makes some (`500`)  of the rows have a different block number.  

```sql
UPDATE address_current_token_balances
SET block_number = block_number - 1
FROM (
     SELECT address_current_token_balances.address_hash,
            address_current_token_balances.token_contract_address_hash
     FROM address_current_token_balances
     LIMIT 500
       ) AS poisoned
WHERE address_current_token_balances.address_hash = poisoned.address_hash AND
      address_current_token_balances.token_contract_address_hash = poisoned.token_contract_address_hash
```

### `poison_all.sql`

**DO NOT RUN THIS ON PROD.**

Purposely makes all of the rows have a different block number.  

```sql
UPDATE address_current_token_balances
SET block_number = block_number - 1
```

## Test Plan

1. Run against Eth Mainnet for a few minutes to generate 100,000-200,000 `address_current_token_balances`.
2. Check that the current code has no incorrect block numbers by running `count.sql` from above.
3. Poison 500 block number with `poison_some.sql` from above.
4. Check that `500` is returned from `count.sql` from above now.
5. Run `update_address_current_token_balances_in_batches.sql` and ensure if also shows `500` block numbers to update and updated.
  ```
  [2019-01-11 16:39:44] [00000] Started at 2019-01-11 22:39:44.936866
  [2019-01-11 16:39:46] [00000] correct_address_current_token_block_numbers TEMP table filled in 00:00:01.5462
  [2019-01-11 16:39:46] [00000] 500 address_current_token_balances to be updated
  [2019-01-11 16:39:46] [00000] -> 500 address current token balances updated.
  [2019-01-11 16:39:46] [00000] -> 500 address tokens block numbers removed from queue.
  [2019-01-11 16:39:46] [00000] -> 0 remaining
  [2019-01-11 16:39:46] [00000] -> 100000 next batch
  [2019-01-11 16:39:46] [00000] -> Estimated time until completion: 00:00:00s
  [2019-01-11 16:39:46] [00000] address_current_token_balances updated in 00:00:00.252498 (00:00:00.000505 per row)
  [2019-01-11 16:39:46] [00000] Ended at 2019-01-11 22:39:46.735909s
  [2019-01-11 16:39:46] [00000] Elapsed time: 00:00:01.799043
  [2019-01-11 16:39:46] completed in 1 s 812 ms
  ```
6. Run `update_address_current_token_balances_in_batches.sql` and ensure if also shows `0` block numbers to update and updated.
  ```
  [2019-01-11 16:39:52] [00000] Started at 2019-01-11 22:39:52.251931
  [2019-01-11 16:39:53] [00000] correct_address_current_token_block_numbers TEMP table filled in 00:00:01.374084
  [2019-01-11 16:39:53] [00000] 0 address_current_token_balances to be updated
  [2019-01-11 16:39:53] [00000] address_current_token_balances updated in 00:00:00.000024 (00:00:00 per row)
  [2019-01-11 16:39:53] [00000] Ended at 2019-01-11 22:39:53.626382s
  [2019-01-11 16:39:53] [00000] Elapsed time: 00:00:01.374451
  [2019-01-11 16:39:53] completed in 1 s 421 ms
  ```
7. Poison all with `poison_all.sql`
8. Run `update_address_current_token_balances_in_batches.sql` and ensure if also shows all block numbers to update and updated.

## Changelog
### Bug Fixes
* Optimize `update_address_current_token_balances_in_batches`
  1. Calculate only the `address_current_token_balances` that need to be updated in the temporary table instead of all `address_current_token_balances`.
  2. Add lot of `RAISE NOTICE` to print
    1. Start Times
    2. End Times
    3. Elapsed Times
    4. Estimate Time of Completion
  3. Only update address_current_token_balances if the block number is less than the temporary tables value, in case other transactions occurred to fix the value while the script is running.

## Upgrading

1. Kill previous `update_address_current_token_balances_in_batches.sql` if it is still running.
2.  Run `update_address_current_token_balances_in_batches.sql` on any machines that need to be updated. 
3. Tell @KronicDeth if it is still estimated to take 58 days to complete.