@amandasposito 2019-01-14
🎨 One thing to consider to decrease the time of this script is to disable the indexes in the `address_current_token_balances` so it does not have to reprocess the index every time there is a write in the table. 

The `checkpoint` clause can be removed as well to see if improves the performance, this clause forces the PostgreSQL to write the things from the memory to the logs, so it causes an I/O operation.

https://www.postgresql.org/docs/9.2/wal-configuration.html
https://dba.stackexchange.com/a/61843